most of the tests implemented and all implemented passing for PatriciaTree

+++ src/Nevermind/Nevermind.Store/Db.cs
@@ -16,7 +16,10 @@ namespace Nevermind.Store
 
         public void Delete(Keccak key)
         {
-            _db.Remove(key);
+            if (_db.ContainsKey(key))
+            {
+                _db.Remove(key);
+            }
         }
 
         // temp
